Added the following functionality: Now users can create routines

+++ Backend/backend/src/main/java/com/your_fitness_journey/backend/Repository/IRoutineRepository.java
@@ -1,9 +1,9 @@
 package com.your_fitness_journey.backend.Repository;
 
-import com.your_fitness_journey.backend.Model.Routine;
+import com.your_fitness_journey.backend.Model.Routines.Routine;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IRoutineRepository extends JpaRepository<Routine, Integer> {
+public interface IRoutineRepository extends JpaRepository<Routine, Long> {
 }
